Add context menu actions(right-click) for the items in the Backup and Browse list views.

+++ browselistwidget.cpp
@@ -20,7 +20,7 @@ void BrowseListWidget::addArchives(QList<ArchivePtr > archives)
     clear();
     foreach (ArchivePtr archive, archives) {
         ArchiveListItem *item = new ArchiveListItem(archive);
-        connect(item, SIGNAL(requestDelete()), this, SLOT(removeItem()));
+        connect(item, SIGNAL(requestDelete()), this, SLOT(removeItems()));
         connect(item, SIGNAL(requestInspect()), this, SLOT(inspectItem()));
         connect(item, SIGNAL(requestRestore()), this, SLOT(restoreItem()));
         this->insertItem(this->count(), item);
@@ -28,15 +28,45 @@ void BrowseListWidget::addArchives(QList<ArchivePtr > archives)
     }
 }
 
-void BrowseListWidget::removeItem()
+void BrowseListWidget::removeItems()
 {
-    ArchiveListItem* archiveItem = qobject_cast<ArchiveListItem*>(sender());
-    if(archiveItem)
+    if(this->selectedItems().count() == 0)
     {
-        emit deleteArchive(archiveItem->archive());
-        QListWidgetItem *item = this->takeItem(this->row(archiveItem));
-        if(item)
-            delete item;
+        // attempt to remove the sender
+        ArchiveListItem* archiveItem = qobject_cast<ArchiveListItem*>(sender());
+        if(archiveItem)
+        {
+            QMessageBox::StandardButton button = QMessageBox::question(this, tr("Confirm action")
+                                                                       , tr("Are you sure you want to delete %1 (this cannot be undone)?").arg(archiveItem->archive()->name));
+            if(button == QMessageBox::Yes)
+            {
+                emit deleteArchive(archiveItem->archive());
+                QListWidgetItem *item = this->takeItem(this->row(archiveItem));
+                if(item)
+                    delete item;
+            }
+        }
+    }
+    else
+    {
+        QMessageBox::StandardButton button = QMessageBox::question(this, tr("Confirm action")
+                                                                   , tr("Are you sure you want to delete %1 selected archives (this cannot be undone)?").arg(this->selectedItems().count()));
+        if(button == QMessageBox::Yes)
+        {
+            foreach (QListWidgetItem *item, this->selectedItems())
+            {
+                if(item->isSelected())
+                {
+                    QListWidgetItem *takenItem = this->takeItem(this->row(item));
+                    if(takenItem)
+                    {
+                        ArchiveListItem* archiveItem = dynamic_cast<ArchiveListItem*>(takenItem);
+                        emit deleteArchive(archiveItem->archive());
+                        delete takenItem;
+                    }
+                }
+            }
+        }
     }
 }
 
@@ -54,38 +84,19 @@ void BrowseListWidget::restoreItem()
 
 }
 
-void BrowseListWidget::removeSelectedItems()
-{
-    QMessageBox::StandardButton button = QMessageBox::question(this, tr("Confirm action")
-                                                               , tr("Are you sure you want to remove %1 selected archives?").arg(this->selectedItems().count()));
-    if(button == QMessageBox::Yes)
-    {
-        foreach (QListWidgetItem *item, this->selectedItems())
-        {
-            if(item->isSelected())
-            {
-                QListWidgetItem *takenItem = this->takeItem(this->row(item));
-                if(takenItem)
-                {
-                    ArchiveListItem* archiveItem = dynamic_cast<ArchiveListItem*>(takenItem);
-                    emit deleteArchive(archiveItem->archive());
-                    delete takenItem;
-                }
-            }
-        }
-    }
-}
-
 void BrowseListWidget::keyReleaseEvent(QKeyEvent *event)
 {
     switch(event->key())
     {
     case Qt::Key_Delete:
     case Qt::Key_Backspace:
-        removeSelectedItems();
+        removeItems();
         break;
     case Qt::Key_Escape:
-        clearSelection();
+        if(this->selectedItems().count() != 0)
+            clearSelection();
+        else
+            QListWidget::keyReleaseEvent(event);
         break;
     default:
         QListWidget::keyReleaseEvent(event);
